Fix for Bug #4: Fire Station gets messed up and turns into Police Station
In MapScanner.java, the doFireStation() function had POLICESTATION as one of its parameters, which is why it would glitch into a broken Police Station after a certain amount of time. I changed this to FIRESTATION and the Fire Station no longer changed.

It mirrors the doPoliceStation() function, which has POLICESTATION as one of its parameters.