Interim Implementation and Progress Report
Progress so far:
I created pixel art for the tool when deselected and selected, and then put those files in the resources folder.

I put the Airtanker in as a tool that can be selected in the toolbar. I set it to cost 10. Pixel art is working correctly.

I duplicated the Bulldozer class and renamed it Airtanker, and then in MicropolisTool had case NEW_BUILDING return new Airtanker. It’s still functionally the same as a Bulldozer, but I can start to revise its behavior in this new java file.