Airtanker Code
This is the code for the Airtanker. I based it off of the Bulldozer and took some of its features for my tool -- the ability to click and drag to define an area and the ability to affect specific tiles.

In the original Bulldozer file, the Bulldozer would detect bulldozeable files and replace them with dirt/river. Airtanker will detect fire tiles and replace them with rubble.

Adding the sound, adding the Airtanker to the toolbar (icon and name), and replicating the Bulldozer was fairly straightforward, but I had trouble figuring out how to detect the fire tiles and only figured out how to get the Airtanker to ignore buildings and just bulldoze rubble.

To modify the tool's behavior so that it would be able to find the fire tiles, Professor Culyba pointed out the TranslatedToolEffect, which allows you to get the tile you're on, and changed the (tile == FIRE) into the range (myTile >= FIRE && myTile <= FIRE+9).